Update user login functionality to utilize custom Next.js API endpoint
Changed functionality of LoginForm (/login) to use a custom Next.js API endpoint to perform user login request to Shopify API.

Customer email/password are signed as a JWT access token and then sent to the endpoint /auth/login. The endpoint then verifies the access token and extracts the email and password, sending them to the Shopify API to be authenticated. The end result is that login functionality persists, and all identifying information is now hidden from the Network tab in web browsers.